--- .storybook/preview.ts
@@ -1,31 +1,0 @@
-import type { Preview } from '@storybook/react'
-
-import { ThemeProvider, CssBaseline } from '@mui/material';
-import { withThemeFromJSXProvider } from '@storybook/addon-themes';
-
-/* TODO: update import for your custom Material UI themes */
-import { lightTheme, darkTheme } from '../path/to/themes';
-
-const preview: Preview = {
-  parameters: {
-    controls: {
-      matchers: {
-       color: /(background|color)$/i,
-       date: /Date$/i,
-      },
-    },
-  },
-
-  decorators: [withThemeFromJSXProvider({
-    GlobalStyles: CssBaseline,
-    Provider: ThemeProvider,
-    themes: {
-      // Provide your custom themes here
-      light: lightTheme,
-      dark: darkTheme,
-    },
-    defaultTheme: 'light',
-  })]
-};
-
-export default preview;
