+++ .storybook/preview.tsx
@@ -1,0 +1,24 @@
+import type { Preview } from '@storybook/react';
+import { ThemeProvider } from '@emotion/react';
+import { theme } from '../src/styles/theme';
+
+const preview: Preview = {
+  parameters: {
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/i,
+      },
+    },
+  },
+
+  decorators: [
+    (Story) => (
+      <ThemeProvider theme={theme}>
+        <Story />
+      </ThemeProvider>
+    ),
+  ],
+};
+
+export default preview;
